+++ DTCD-DevPanel-Console/src/Plugin.Meta.js
@@ -1,7 +1,7 @@
 export default {
   type: 'extension',
   target: 'DevPanel',
-  name: 'DevPanelConsole',
+  name: 'DevPanel_Console',
   title: 'Консоль',
   version: '0.1.0',
   withDependencies: true,