+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# **CHANGELOG**
+# Changelog
 
 All notable changes to this project will be documented in this file.
 
@@ -8,11 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
-
+- plugin build process
 - name of plugin from DevPanel-Console to DevPanel_Console
 - the way autcomplete works
 - makefile
 
 ### Removed
-
 - autocomplete hardcode file